+++ src/com/mac/rltut/engine/graphics/Font.java
@@ -8,9 +8,7 @@ import com.esotericsoftware.minlog.Log;
  * Created by Matt on 25/06/2017 at 10:55 AM.
  */
 public class Font extends Spritesheet{
-
-    public static final Font font = new Font("anikki_square", 8, 8);
-
+    
     private int charWidth, charHeight;
     private Sprite[] chars;