+++ src/com/mac/rltut/engine/graphics/Spritesheet.java
@@ -7,16 +7,16 @@ import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.HashMap;
 
 /**
  * Project: complete-rltut
  * PC
  * Created by Matt on 25/06/2017 at 10:55 AM.
  */
-public class Spritesheet extends Bitmap{
+public class Spritesheet extends Bitmap {
 
-    public static final Spritesheet textures = new Spritesheet("minirogue-all");
-    public static final Spritesheet fog = new Spritesheet("fog");
+    public static HashMap<String, Spritesheet> sheets = new HashMap<>();
 
     private String name;
 
@@ -25,9 +25,9 @@ public class Spritesheet extends Bitmap{
         load();
     }
 
-    public Sprite cutSprite(int xp, int yp, int w, int h, int tileSize){
-        xp *= tileSize;
-        yp *= tileSize;
+    public Sprite cutSprite(int xp, int yp, int w, int h, int tileIndexSize){
+        xp *= tileIndexSize;
+        yp *= tileIndexSize;
         int[] spritePixels = new int[w * h];
         for(int y = 0; y < h; y++){
             int ya = y + yp;
@@ -47,9 +47,20 @@ public class Spritesheet extends Bitmap{
             height = image.getHeight();
             pixels = new int[width * height];
             image.getRGB(0, 0, width, height, pixels, 0, width);
-            Log.debug("Loaded spritesheet [" + name + "]");
         } catch (IOException e) {
             e.printStackTrace();
         }
+    }
+
+    public static Spritesheet get(String name) {
+        if(!sheets.containsKey(name)){
+            Log.warn("Spritesheet [" + name + "] does not exist.");
+            return null;
+        }
+        return sheets.get(name);
+    }
+    
+    public static void add(String name, Spritesheet sheet){
+        sheets.put(name, sheet);
     }
 }