+++ src/com/mac/rltut/game/world/tile/Tile.java
@@ -2,6 +2,8 @@ package com.mac.rltut.game.world.tile;
 
 import com.mac.rltut.engine.graphics.Sprite;
 
+import java.util.HashMap;
+
 /**
  * Project: complete-rltut
  * PC
@@ -11,49 +13,38 @@ public abstract class Tile {
     
     private static Tile[] tiles = new Tile[256]; 
     private static int currentId = 0;
-    
-    public static final Tile empty = new EmptyTile(Sprite.empty, "empty");
-    public static final Tile floor = new EmptyTile(Sprite.empty, "floor");
-    
+
+    public static final Tile empty = new EmptyTile(Sprite.get("empty"), "empty");
+    public static final Tile floor = new EmptyTile(Sprite.get("empty"), "floor");
+
     //TILES
-    public static final Tile wallTopRed = new SolidTile(Sprite.wallTopRed, "wall");
-    public static final Tile wallTopBlue = new SolidTile(Sprite.wallTopBlue, "wall");
-    public static final Tile wallSide = new SolidTile(Sprite.wallSide, "wall");
-    public static final Tile stairDown = new EmptyTile(Sprite.stairDown, "stair");
-    public static final Tile stairUp = new EmptyTile(Sprite.stairUp, "stair");
-    public static final Tile grassYellow = new EmptyTile(Sprite.grassYellow, "grass");
-    public static final Tile grassGreen = new EmptyTile(Sprite.grassGreen, "grass");
-    public static final Tile grassBlue = new EmptyTile(Sprite.grassBlue, "grass");
-    public static final Tile grassPurple = new EmptyTile(Sprite.grassPurple, "grass");
-    
-    public static final Tile treeDeciduous = new SolidTile(Sprite.treeDeciduous, "tree");
-    public static final Tile treeConifer = new SolidTile(Sprite.treeConifer, "tree");
+    public static final Tile wallTopRed = new SolidTile(Sprite.get("wallTopRed"), "wall");
+    public static final Tile wallTopBlue = new SolidTile(Sprite.get("wallTopBlue"), "wall");
+    public static final Tile wallSide = new SolidTile(Sprite.get("wallSide"), "wall");
+    public static final Tile stairDown = new EmptyTile(Sprite.get("stairDown"), "stair");
+    public static final Tile stairUp = new EmptyTile(Sprite.get("stairUp"), "stair");
+    public static final Tile grassYellow = new EmptyTile(Sprite.get("grassYellow"), "grass");
+    public static final Tile grassGreen = new EmptyTile(Sprite.get("grassGreen"), "grass");
+    public static final Tile grassBlue = new EmptyTile(Sprite.get("grassBlue"), "grass");
+    public static final Tile grassPurple = new EmptyTile(Sprite.get("grassPurple"), "grass");
 
-    public static final Tile lavaLight = new ImpassableTile(Sprite.lavaYellow, "lava");
-    public static final Tile lavaDark = new ImpassableTile(Sprite.lavaBrown, "lava");
-    
-    public static final Tile waterFoul = new ImpassableTile(Sprite.waterPurple, "water");
-    public static final Tile waterDirty = new ImpassableTile(Sprite.waterBrown, "water");
-    public static final Tile waterBlue = new ImpassableTile(Sprite.waterBlue, "water");
-    public static final Tile waterLilypad = new ImpassableTile(Sprite.waterLilypad, "water");
-    public static final Tile waterBonesFoul1 = new ImpassableTile(Sprite.waterBonesPurple1, "water");
-    public static final Tile waterBonesFoul2 = new ImpassableTile(Sprite.waterBonesPurple2, "water");
-    public static final Tile waterBonesDirty1 = new ImpassableTile(Sprite.waterBonesBrown1, "water");
-    public static final Tile waterBonesDirty2 = new ImpassableTile(Sprite.waterBonesBrown2, "water");
-    
-    public static final Tile blood = new EmptyTile(Sprite.blood, "blood");
+    public static final Tile treeDeciduous = new SolidTile(Sprite.get("treeDeciduous"), "tree");
+    public static final Tile treeConifer = new SolidTile(Sprite.get("treeConifer"), "tree");
 
-    public static final Tile doorSilver = new EmptyTile(Sprite.doorSilver, "door");
-    public static final Tile doorGold = new EmptyTile(Sprite.doorGold, "door");
-    public static final Tile doorGreen = new EmptyTile(Sprite.doorGreen, "door");
-    public static final Tile doorPurple = new EmptyTile(Sprite.doorPurple, "door");
-    
-    public static final Tile chestGold = new ChestTile(Sprite.chestGoldClosed, Sprite.chestGoldOpen, "chest");
-    public static final Tile chestSilver = new ChestTile(Sprite.chestSilverClosed, Sprite.chestSilverOpen, "chest");
-    
-    public static final Tile mushroom = new EmptyTile(Sprite.mushroom, "mushroom");
-    
-    
+    public static final Tile waterFoul = new ImpassableTile(Sprite.get("waterPurple"), "water");
+    public static final Tile waterDirty = new ImpassableTile(Sprite.get("waterBrown"), "water");
+    public static final Tile waterBlue = new ImpassableTile(Sprite.get("waterBlue"), "water");
+    public static final Tile waterLilypad = new ImpassableTile(Sprite.get("waterLilypad"), "water");
+    public static final Tile waterBonesFoul1 = new ImpassableTile(Sprite.get("waterBonesPurple1"), "water");
+    public static final Tile waterBonesFoul2 = new ImpassableTile(Sprite.get("waterBonesPurple2"), "water");
+    public static final Tile waterBonesDirty1 = new ImpassableTile(Sprite.get("waterBonesBrown1"), "water");
+    public static final Tile waterBonesDirty2 = new ImpassableTile(Sprite.get("waterBonesBrown2"), "water");
+
+    public static final Tile chestGold = new ChestTile(Sprite.get("chestGoldClosed"), Sprite.get("chestGoldOpen"), "chest");
+    public static final Tile chestSilver = new ChestTile(Sprite.get("chestSilverClosed"), Sprite.get("chestSilverOpen"), "chest");
+
+    public static final Tile mushroom = new EmptyTile(Sprite.get("mushroom"), "mushroom");
+        
     public final byte id;
     private Sprite sprite;
     private String type;